requirements.txt
flask
flask-socketio
pyautogui
mss
pillow
python-socketio
werkzeug